* Goal
    * Provide the capability within SOILWAT2 to run many simulations at once, i.e., run simulations in parallel
    * This should provide a speedup over running simulations sequentially in SWNC/SWNETCDF mode

* New mode - `SWMPI`
* Add capabilities to use the library OpenMPI which allows us to
    * Create many processes to work on a simulation domain
    * Divide the work among these many processes
* Use the parallel netCDF library for parallel I/O interactions with files
* Divide the spawned processes into "compute" and "I/O" processes where
    * I/O Processes - Read and write data from/to files and provide information to compute processes
    * Compute Processes - Get input, simulate, and send out simulated output to I/O processes

+++ include/rands.h
@@ -12,6 +12,7 @@
 
 #include "include/SW_datastructs.h" // for LOG_INFO
 #include "include/SW_Defines.h"     // for sw_random_t
+#include <stddef.h>                 // for size_t
 
 #ifdef __cplusplus
 extern "C" {
@@ -27,9 +28,7 @@ typedef long RandListType;
 /* =================================================== */
 /*             Global Function Declarations            */
 /* --------------------------------------------------- */
-void RandSeed(
-    unsigned long initstate, unsigned long initseq, sw_random_t *pcg_rng
-);
+void RandSeed(size_t initstate, size_t initseq, sw_random_t *pcg_rng);
 
 double RandUni(sw_random_t *pcg_rng);
